testing (map): move isPulsating and timerPing to +layout.svelte
Does not work yet
